+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod list;
 pub mod locale;
 pub mod number;
 pub mod ordinal;
+pub mod percent;
 pub mod prelude;
 #[cfg(feature = "time")]
 #[cfg_attr(docsrs, doc(cfg(feature = "time")))]
@@ -29,4 +30,5 @@ pub use error::{DurationConversionError, NegativeDurationError};
 pub use list::{list, list_with, ListDisplay, ListOptions};
 pub use number::{number, number_with, NumberDisplay, NumberOptions};
 pub use ordinal::{ordinal, ordinal_with, OrdinalDisplay, OrdinalLike};
+pub use percent::{percent, percent_with, PercentDisplay, PercentLike, PercentOptions};
 pub use traits::Humanize;